have tto move props and params logic to client side
