+++ web/app/(portal)/teams/[teamId]/apps/[appId]/actions/page.tsx
@@ -1,58 +1,9 @@
 import { generateMetaTitle } from "@/lib/genarate-title";
-import { urls } from "@/lib/urls";
-import { fetchAppEnvCached } from "@/scenes/Portal/Teams/TeamId/Apps/AppId/layout/server/fetch-app-env";
 import { ActionsPage } from "@/scenes/Portal/Teams/TeamId/Apps/AppId/Actions/page";
 import { Metadata } from "next";
-import { redirect } from "next/navigation";
-import { AppLayoutRouteParams } from "../layout-params";
 
 export const metadata: Metadata = {
   title: generateMetaTitle({ left: "Incognito actions" }),
 };
 
-type Props = {
-  params: AppLayoutRouteParams;
-  searchParams: Promise<Record<string, string>>;
-};
-
-export default async function Page(props: Props) {
-  const params = await props.params;
-  const searchParams = await props.searchParams;
-  const { action, app } = await fetchAppEnvCached(params.appId);
-  const hasLegacyActions = (action?.length ?? 0) > 0;
-  const hasRpRegistration = (app?.[0]?.rp_registration?.length ?? 0) > 0;
-  const showLegacyList = searchParams.legacy === "true" && hasLegacyActions;
-
-  if (showLegacyList) {
-    return <ActionsPage {...props} />;
-  }
-
-  if (hasLegacyActions && !hasRpRegistration) {
-    const legacyUrl = urls.legacyActions({
-      team_id: params.teamId,
-      app_id: params.appId,
-    });
-    redirect(
-      searchParams.createAction === "true"
-        ? `${legacyUrl}&createAction=true`
-        : legacyUrl,
-    );
-  }
-
-  if (!hasRpRegistration) {
-    redirect(
-      urls.enableWorldId4({ team_id: params.teamId, app_id: params.appId }),
-    );
-  }
-
-  const worldIdActionsUrl = urls.worldIdActions({
-    team_id: params.teamId,
-    app_id: params.appId,
-  });
-
-  redirect(
-    searchParams.createAction === "true"
-      ? `${worldIdActionsUrl}?createAction=true`
-      : worldIdActionsUrl,
-  );
-}
+export default ActionsPage;
