+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Actions/page/graphql/client/app.graphql
@@ -2,8 +2,10 @@ query GetApp($app_id: String!) {
   app: app_by_pk(id: $app_id) {
     id
     engine
+    rp_registration {
+      rp_id
+    }
     app_metadata {
-      id
       name
     }
   }
